move runtime/execution.ts to runtime/common/source_map.ts (#1240)
